♻️ refactor: 리액트 쿼리 프로바이더 앱 라우터에 맞게 수정 #1
